E2fsprogs 1.43.4 release fixup
The orig tar.gz that was uploaded to e2fsprogs 1.43.4-1 was not the
final release tarball; it had an outdated version of e2fsprogs.lsm
and e2fsprogs.spec.  This doesn't affect generated binary packages of
e2fsprogs at all, but let's correct it to make it easier generate
updated packages.

Gbp-Pq: Name e2fsprogs-1.43.4-release-fixup